requirements.txt
 Flask==2.2.5
playwright==1.40.0
pandas==2.2.2
pdfplumber==0.8.0
PyMuPDF==1.22.0
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
openai==1.0.0

gunicorn==20.1.0
